fix build errors, styling, debugging login
