+++ packages/blockchain-tools/src/chains/BlockchainInterfaces.ts
@@ -1,12 +1,7 @@
-import appConfig from "@cfce/app-config"
 import type {
-  ChainSlugs,
   ClientInterfaces,
-  Network,
-  NetworkConfig,
   ServerInterfaces,
 } from "@cfce/types"
-import { getNetwork } from "@stellar/freighter-api"
 import { keys as _keys } from "lodash"
 import { get as _get } from "lodash"
 import type { Interface } from "../interfaces"
@@ -20,20 +15,11 @@ import Web3Server from "../interfaces/Web3Server"
 import XrplServer from "../interfaces/XrplServer"
 import XummClient from "../interfaces/XummClient"
 import type InterfaceBaseClass from "./InterfaceBaseClass"
-import chainConfiguration from "./chainConfig"
 
 type ChainClasses<ClientClass = Interface, ServerClass = Interface> = {
   client: Partial<Record<ClientInterfaces, InterfaceBaseClass>>
   server: InterfaceBaseClass
 }
-
-/**
- * Get the network config for the chain, using appConfig.chainDefaults.network
- * @param slug - The chain slug
- * @returns The network config object
- */
-const getNetworkForChain = (slug: ChainSlugs): NetworkConfig =>
-  chainConfiguration[slug].networks[appConfig.chainDefaults.network]
 
 const BlockchainClientInterfaces: Record<ClientInterfaces, InterfaceBaseClass> =
   {
@@ -53,7 +39,6 @@ const BlockchainServerInterfaces = {
 } satisfies Record<ServerInterfaces, InterfaceBaseClass>
 
 export {
-  getNetworkForChain,
   BlockchainClientInterfaces,
   BlockchainServerInterfaces,
 }
