+++ packages/blockchain-tools/src/chains/index.ts
@@ -6,3 +6,4 @@ export {
 export { default as chainConfig } from "./chainConfig"
 export * from "./chainConfig"
 export * from "./useWalletClient"
+export * from './utils'
