+++ packages/blockchain-tools/src/interfaces/StellarServer.ts
@@ -13,10 +13,10 @@ import type {
   // Keypair,
   // TransactionBuilder,
 } from "@stellar/stellar-sdk"
-import { chainConfig } from "../chains"
-import { getNetworkForChain } from "../chains/BlockchainInterfaces"
 // import Stellar, { StellarNetworks } from "./common"
 import InterfaceBaseClass from "../chains/InterfaceBaseClass"
+import chainConfig from "../chains/chainConfig"
+import { getNetworkForChain } from "../chains/utils"
 import Contract721 from "../contracts/soroban/nft721/server"
 
 export default class StellarServer extends InterfaceBaseClass {
