+++ packages/blockchain-tools/src/interfaces/Web3Server.ts
@@ -6,9 +6,9 @@ import type {
 } from "@cfce/types"
 import _get from "lodash/get"
 import Web3 from "web3"
-import { chainConfig } from "../chains"
-import { getNetworkForChain } from "../chains/BlockchainInterfaces"
 import InterfaceBaseClass from "../chains/InterfaceBaseClass"
+import chainConfig from "../chains/chainConfig"
+import { getNetworkForChain } from "../chains/utils"
 import Abi721base from "../contracts/solidity/erc721/erc721base-abi.json" // must pass tokenid
 import Abi721inc from "../contracts/solidity/erc721/erc721inc-abi.json" // autoincrements tokenid
 import Abi1155 from "../contracts/solidity/erc1155/erc1155-abi.json"
